Added support for named place holders. This probably breaks server-side prepare but this is a good stopping place. This also breaks the LIMIT placeholder detection code thereby breaking some tests in the limit.t.
Also removed do() from msyql.xs as it is no longer needed -- there is no need
to set up the execute statement anymore, because the info is
now kept in imp_dbh.

t/insert.t needed to be updated because it was calling execute with an
unbound place holder which now thows an error.


git-svn-id: https://svn.perl.org/modules/DBD-mysql/branches/Dev-3_0@1168 50811bd7-b8ce-0310-adc1-d9db26280581