Made test suite more robust
Previously, when a database existed but the test user did not have
permission, the tests failed. Unfortunately, on some CPAN smokers
this was the case.

Now we skip those tests and make it clear there is no database
connection.

Also, two newer tests did not properly bail out when there was no
database available.