Improve test dependencies
t/87async.t has some kind of test to skip if Test::Deep is not
installed but it had a use statement for the module; so this
did not work - mentioned by Vladimir Marek in
https://rt.cpan.org/Ticket/Display.html?id=112072

However since Test::Deep is correctly specified as a test
dependency in Makefile.PL and thus in META.(json|yaml) the test
makes no sense and I removed it.

Also, Time::HiRes is used by this test; it's a core module but
was not specified in Makefile.PL - this can cause problems on
platforms that do not ship with all core perl modules.
(I'm looking at YOU, RHEL6!)