Extend test t/40server_prepare_crash.t for libmysqlclient.so crash
MySQL library libmysqlclient.so in versions prior to MySQL 5.6 release has
defect and returns dangling pointers for all statements after closing
connection with MySQL server. Moreover libmysqlclient.so itself when
freeing statement is accessing that (already freed) dangling pointer (which
represent connection with MySQL server) and that leads to memory
corruption.

Because in perl it is not guaranteed that DESTROY of $sth (close statement)
will be called before DESTROY of $dbh (close connection to MySQL server and
free structures) that memory corruption can cause crash of DBD::mysql and
perl interpreter.

This patch extend t/40server_prepare_crash.t test and check that MySQL
library libmysqlclient.so correctly invalidate pointer to NULL and so
closing and freeing statement will be without accessing already freed
memory. It used new "mysql_sock" $sth attribute for that purpose.

Fix for libmysqlclient.so is in MySQL commit:
https://github.com/mysql/mysql-server/commit/4797ea0b772d5f4c5889bc552424132806f46e93

For more information is in reported MySQL bugs:
https://bugs.mysql.com/bug.php?id=70429
https://bugs.mysql.com/bug.php?id=63363