Fix t/05dbcreate.t when configured with --testsocket option
The 05dbcreate.t test tries to connect to database without a database
name specified. It does this by taking the DSN being used and removing
the database name. The problem was that the code didn't expect any
additional parameters separated by ';' (like database socket) and so
removed that part of DSN too.