fix detection of client version on MariaDB 10.6.x
For mysql_get_client_version(), MariaDB 10.6.2+ reports the version of
the mariadb-connector-c library rather than the server version:
https://jira.mariadb.org/browse/CONC-509

The mariadb-connector-c library versioning does not track the versioning
of the main MariaDB client/server suite, so the change broke the client
version comparisons here in DBD-mysql. The previous behavior of
reporting the _server_ version probably mostly worked by accident, since
server and client versions tend to be roughly aligned in user
environments.

To fix this, change DBD-mysql to use a function which is documented to
return the appropriate client version information.
https://mariadb.com/kb/en/mariadb_get_infov/

NOTE: upstream documentation on mariadb_get_infov is incorrect. The
call modifies a size_t, not a unsigned int.

This fixes:
https://github.com/perl5-dbi/DBD-mysql/issues/333
...and is a minor revision of @bviviano's original fix therein.