fix detection of client version on MariaDB 10.6.x
For mysql_get_client_version(), MariaDB 10.6.2+ reports the version of
the mariadb-connector-c library rather than the client version:
https://jira.mariadb.org/browse/CONC-509

There is some confusion in that ticket for client vs. server version,
and they may have been the same, but the code really did report the
client version. See this explanation by @pali for more details:
https://github.com/perl5-dbi/DBD-mysql/pull/339#issuecomment-1093274153

To fix this, change DBD-mysql to use a function which is documented to
return the appropriate client version information.
https://mariadb.com/kb/en/mariadb_get_infov/

This fixes:
https://github.com/perl5-dbi/DBD-mysql/issues/333
...and is a minor revision of @bviviano's original fix therein.

Thanks also to @toddr for spotting the variable type discrepancy.