Add tests for UTF8 characters in table definition and errors
The tests are taken from DBD::MariaDB and adopted to DBD::mysql.

Testing is provided on Japanese and Chinese characters.

There are two test files:
* one for testing of utf8/utf8mb4 identifiers
* one for testing of utf8 error messages

There are three runs of all tests in both test files:
* one without mysql_enable_utf8 option
* one with enabled mysql_enable_utf8 option
* one with enabled mysql_enable_utf8mb4 option

There are XXX comments for situations which not follow
mysql_enable_utf8/mb4 option.